View: delivery-view & movie-materials showing correct informations
