+++ libs/material/src/lib/delivery/delivery-form/delivery-form.component.scss
@@ -1,4 +1,3 @@
-
 mat-sidenav-container {
   height: 100%;
   min-height: 400px;
