--- libs/material/src/lib/delivery/delivery-materials/delivery-materials.component.html
@@ -1,41 +1,0 @@
-<button mat-raised-button color="warn" (click)="goBack()">Back</button>
-<section fxLayout="row" fxLayoutGap="5%">
-  <article fxFlex="30">
-    <mat-card>
-      <mat-card-content *ngIf="(progressionValue$ | async) as value" fxLayout="column" fxLayoutAlign="center center">
-        <h2 class="example-h2">PROGRESSION</h2>
-        <mat-progress-spinner
-          class="example-margin"
-          color="primary"
-          mode="determinate"
-          [value]="value"
-          diameter="200"
-          strokeWidth="13"
-        >
-        </mat-progress-spinner>
-        <div class ="progress-value">
-            {{value}} %
-        </div>
-        <p><i>La progression en %age de la delivery</i></p>
-      </mat-card-content>
-    </mat-card>
-  </article>
-  <article fxFlex="70" fxLayout="column">
-    <mat-card *ngFor="let category of (materials$ | async | keyvalue)">
-      <mat-card-title>{{ category.key | uppercase }}</mat-card-title>
-      <mat-card-content *ngFor="let material of category.value">
-        <span
-          ><h4 mat-line>
-            <mat-checkbox [(ngModel)]="material.delivered" [disabled]="true"
-              ><b>{{ material.value }}</b></mat-checkbox
-            >
-          </h4></span
-        >
-        <p mat-line class="description-text">
-          <i>{{ material.description }}</i>
-        </p>
-        <mat-divider></mat-divider>
-      </mat-card-content>
-    </mat-card>
-  </article>
-</section>
