--- libs/material/src/lib/delivery/delivery-materials/delivery-materials.component.scss
@@ -1,34 +1,0 @@
-::ng-deep .mat-list .mat-list-item .mat-line{
-  word-wrap: break-word;
-  white-space: pre-wrap;
-}
-
-.mat-card {
-  margin-bottom: 2rem;
-}
-
-mat-card-title {
-  font-weight: bolder;
-}
-
-.mat-divider {
-  width: 80%;
-  text-align: center;
-}
-
-h4 {
-  color: #5E5E5E;
-}
-
-p {
-  font-size: 90%;
-  margin-bottom: 1rem;
-}
-
-.progress-value {
-  position: relative;
-  top: -120px;
-  color: #3F51B4;
-  font-weight: bolder;
-  font-size: 2.5rem;
-}
