+++ libs/material/src/lib/delivery/delivery-view/delivery-view.component.html
@@ -1,36 +1,59 @@
-<mat-sidenav-container>
-  <!-- Sidenav -->
-  <mat-sidenav [opened]="true" mode="side">
-    <material-list *ngIf="(delivery$ | async) as delivery" [template]="delivery" name="Delivery"></material-list>
-  </mat-sidenav>
-  <!-- Content -->
-  <mat-sidenav-content>
-    <button (click)="addForm('')" color="primary" mat-icon-button matRipple>
-      <mat-icon class="icon-add">add_circle</mat-icon>
-      Add new material
-    </button>
-    <div *ngIf="(form$ | async) as form">
-      <material-form *ngIf="form.category === ''" (material)="addMaterial($event)"></material-form>
-    </div>
-    <mat-card *ngFor="let materials of (delivery$ | async | keyvalue)" id="{{ materials.key }}">
-      <mat-card-title class="container" fxLayout="row" fxLayoutAlign="start center">
-        <h4 fxFlex="60">Category : {{ materials.key }}</h4>
-        <button fxFlex="40" (click)="addForm(materials.key)" color="primary" mat-icon-button matRipple>
-          <mat-icon class="icon-add">add_circle</mat-icon>
-          Add new material with this category
-        </button>
-      </mat-card-title>
-      <mat-card-content>
-        <div cdkDrag *ngFor="let material of materials.value">
-          <material-item [material]="material" (isDeleted)="deleteMaterial(material)"></material-item>
+<button mat-raised-button color="warn" (click)="goBack()">Back</button>
+<section fxLayout="row" fxLayoutGap="5%">
+  <article class="progress-card" fxFlex="30">
+    <mat-card>
+      <mat-card-content *ngIf="(progressionValue$ | async) as value else empty" fxLayout="column" fxLayoutAlign="center center">
+        <h2 class="example-h2">PROGRESSION</h2>
+        <mat-progress-spinner
+          class="example-margin"
+          color="primary"
+          mode="determinate"
+          [value]="value"
+          diameter="200"
+          strokeWidth="13"
+        >
+        </mat-progress-spinner>
+        <div class ="progress-value">
+            {{value}} %
         </div>
-        <div *ngIf="(form$ | async) as form">
-          <material-form
-            *ngIf="form.category === materials.key"
-            (material)="addMaterial($event)"
-          ></material-form>
-        </div>
+        <p><i>La progression en %age des deliveries de la movie</i></p>
+      </mat-card-content>
+      <ng-template #empty>
+        <mat-card-content fxLayout="column" fxLayoutAlign="center center">
+          <h2 class="example-h2">PROGRESSION</h2>
+          <mat-progress-spinner
+            class="example-margin"
+            color="primary"
+            mode="determinate"
+            value="0"
+            diameter="200"
+            strokeWidth="15"
+          >
+          </mat-progress-spinner>
+          <div class ="progress-value">
+              0 %
+          </div>
+          <p><i>La progression en %age des deliveries de la movie</i></p>
+        </mat-card-content>
+      </ng-template>
+    </mat-card>
+  </article>
+  <article fxFlex="70" fxLayout="column">
+    <mat-card *ngFor="let category of (materials$ | async | keyvalue)">
+      <mat-card-title>{{ category.key | uppercase }}</mat-card-title>
+      <mat-card-content *ngFor="let material of category.value">
+        <span
+          ><h4 mat-line>
+            <mat-checkbox [(ngModel)]="material.delivered" [disabled]="true"
+              ><b>{{ material.value }}</b></mat-checkbox
+            >
+          </h4></span
+        >
+        <p mat-line class="description-text">
+          <i>{{ material.description }}</i>
+        </p>
+        <mat-divider></mat-divider>
       </mat-card-content>
     </mat-card>
-  </mat-sidenav-content>
-</mat-sidenav-container>
+  </article>
+</section>
