+++ libs/material/src/lib/delivery/delivery-view/delivery-view.component.scss
@@ -1,25 +1,34 @@
-mat-sidenav-container {
-  height: 100%;
-  min-height: 400px;
+::ng-deep .mat-list .mat-list-item .mat-line{
+  word-wrap: break-word;
+  white-space: pre-wrap;
 }
 
-mat-sidenav {
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 100%;
-  width: 250px;
-}
-
-mat-sidenav-content {
-  height: 100%;
-  overflow: hidden;
-}
-
-mat-card {
-  margin: 5% 5%;
+.mat-card {
+  margin-bottom: 2rem;
 }
 
 mat-card-title {
-  border-bottom: 1px solid black;
+  font-weight: bolder;
+}
+
+.mat-divider {
+  width: 80%;
+  text-align: center;
+}
+
+h4 {
+  color: #5E5E5E;
+}
+
+p {
+  font-size: 90%;
+  margin-bottom: 1rem;
+}
+
+.progress-value {
+  position: relative;
+  top: -120px;
+  color: #3F51B4;
+  font-weight: bolder;
+  font-size: 2.5rem;
 }
